requirements.txt
fastapi
uvicorn

sentence-transformers==3.0.1
faiss-cpu==1.8.0.post1
langdetect==1.0.9
huggingface_hub==0.24.5

llama-cpp-python==0.2.90
